Rename relbench.py to _relbench.py
